+++ src/TorchSharp/NN/Convolution/Conv3D.cs
@@ -61,92 +61,92 @@ namespace TorchSharp
             /// <summary>
             /// Applies a 3D convolution over an input signal composed of several input planes
             /// </summary>
-            /// <param name="inputChannel">Number of channels in the input image</param>
-            /// <param name="outputChannel">Number of channels produced by the convolution</param>
+            /// <param name="in_channels">Number of channels in the input image</param>
+            /// <param name="out_channels">Number of channels produced by the convolution</param>
             /// <param name="kernelSize">Size of the convolving kernel</param>
             /// <param name="stride">Stride of the convolution. Default: 1</param>
             /// <param name="padding">Zero-padding added to both sides of the input. Default: 0</param>
             /// <param name="dilation">Spacing between kernel elements. Default: 1</param>
-            /// <param name="paddingMode">'zeros', 'reflect', 'replicate' or 'circular'. Default: 'zeros'</param>
+            /// <param name="padding_mode">'zeros', 'reflect', 'replicate' or 'circular'. Default: 'zeros'</param>
             /// <param name="groups">Number of blocked connections from input channels to output channels. Default: 1</param>
             /// <param name="bias">If true, adds a learnable bias to the output. Default: true</param>
             /// <param name="device">The desired device of the parameters and buffers in this module</param>
             /// <param name="dtype">The desired floating point or complex dtype of the parameters and buffers in this module</param>
-            public static Conv3d Conv3d(long inputChannel, long outputChannel, long kernelSize, long stride = 1, long padding = 0, long dilation = 1, PaddingModes paddingMode = PaddingModes.Zeros, long groups = 1, bool bias = true, Device? device = null, ScalarType? dtype = null)
+            public static Conv3d Conv3d(long in_channels, long out_channels, long kernelSize, long stride = 1, long padding = 0, long dilation = 1, PaddingModes padding_mode = PaddingModes.Zeros, long groups = 1, bool bias = true, Device? device = null, ScalarType? dtype = null)
             {
-                var res = THSNN_Conv3d_ctor(inputChannel, outputChannel, kernelSize, stride, padding, dilation, (long)paddingMode, groups, bias, out var boxedHandle);
+                var res = THSNN_Conv3d_ctor(in_channels, out_channels, kernelSize, stride, padding, dilation, (long)padding_mode, groups, bias, out var boxedHandle);
                 if (res == IntPtr.Zero) { torch.CheckForErrors(); }
-                return new Conv3d(res, boxedHandle, inputChannel).MoveModule<Conv3d>(device, dtype);
+                return new Conv3d(res, boxedHandle, in_channels).MoveModule<Conv3d>(device, dtype);
             }
 
             /// <summary>
             /// Applies a 3D convolution over an input signal composed of several input planes
             /// </summary>
-            /// <param name="inputChannel">Number of channels in the input image</param>
-            /// <param name="outputChannel">Number of channels produced by the convolution</param>
+            /// <param name="in_channels">Number of channels in the input image</param>
+            /// <param name="out_channels">Number of channels produced by the convolution</param>
             /// <param name="kernelSize">Size of the convolving kernel</param>
             /// <param name="stride">Stride of the convolution. Default: (1,1,1)</param>
             /// <param name="padding">Zero-padding added to both sides of the input. Default: (0,0,0)</param>
             /// <param name="dilation">Spacing between kernel elements. Default: (1,1,1)</param>
-            /// <param name="paddingMode">'zeros', 'reflect', 'replicate' or 'circular'. Default: 'zeros'</param>
+            /// <param name="padding_mode">'zeros', 'reflect', 'replicate' or 'circular'. Default: 'zeros'</param>
             /// <param name="groups">Number of blocked connections from input channels to output channels. Default: 1</param>
             /// <param name="bias">If true, adds a learnable bias to the output. Default: true</param>
             /// <param name="device">The desired device of the parameters and buffers in this module</param>
             /// <param name="dtype">The desired floating point or complex dtype of the parameters and buffers in this module</param>
-            public static Conv3d Conv3d(long inputChannel, long outputChannel, (long, long, long) kernelSize, (long, long, long)? stride = null, (long, long, long)? padding = null, (long, long, long)? dilation = null, PaddingModes paddingMode = PaddingModes.Zeros, long groups = 1, bool bias = true, Device? device = null, ScalarType? dtype = null)
+            public static Conv3d Conv3d(long in_channels, long out_channels, (long, long, long) kernelSize, (long, long, long)? stride = null, (long, long, long)? padding = null, (long, long, long)? dilation = null, PaddingModes padding_mode = PaddingModes.Zeros, long groups = 1, bool bias = true, Device? device = null, ScalarType? dtype = null)
             {
                 if (stride == null) stride = (1, 1, 1);
                 if (padding == null) padding = (0, 0, 0);
                 if (dilation == null) dilation = (1, 1, 1);
 
-                var res = THSNN_Conv3d_ctor_1(inputChannel, outputChannel, kernelSize.Item1, kernelSize.Item2, kernelSize.Item3, stride.Value.Item1, stride.Value.Item2, stride.Value.Item3, padding.Value.Item1, padding.Value.Item2, padding.Value.Item3, dilation.Value.Item1, dilation.Value.Item2, dilation.Value.Item3, (long)paddingMode, groups, bias, out var boxedHandle);
+                var res = THSNN_Conv3d_ctor_1(in_channels, out_channels, kernelSize.Item1, kernelSize.Item2, kernelSize.Item3, stride.Value.Item1, stride.Value.Item2, stride.Value.Item3, padding.Value.Item1, padding.Value.Item2, padding.Value.Item3, dilation.Value.Item1, dilation.Value.Item2, dilation.Value.Item3, (long)padding_mode, groups, bias, out var boxedHandle);
                 if (res == IntPtr.Zero) { torch.CheckForErrors(); }
-                return new Conv3d(res, boxedHandle, inputChannel).MoveModule<Conv3d>(device, dtype);
+                return new Conv3d(res, boxedHandle, in_channels).MoveModule<Conv3d>(device, dtype);
             }
 
             /// <summary>
             /// Applies a 3D convolution over an input signal composed of several input planes
             /// </summary>
-            /// <param name="inputChannel">Number of channels in the input image</param>
-            /// <param name="outputChannel">Number of channels produced by the convolution</param>
+            /// <param name="in_channels">Number of channels in the input image</param>
+            /// <param name="out_channels">Number of channels produced by the convolution</param>
             /// <param name="kernelSize">Size of the convolving kernel</param>
             /// <param name="stride">Stride of the convolution. Default: 1</param>
             /// <param name="padding">Zero-padding added to both sides of the input. padding=Valid is the same as no padding. padding=Same pads the input so the output has the shape as the input. </param>
             /// <param name="dilation">Spacing between kernel elements. Default: 1</param>
-            /// <param name="paddingMode">'zeros', 'reflect', 'replicate' or 'circular'. Default: 'zeros'</param>
+            /// <param name="padding_mode">'zeros', 'reflect', 'replicate' or 'circular'. Default: 'zeros'</param>
             /// <param name="groups">Number of blocked connections from input channels to output channels. Default: 1</param>
             /// <param name="bias">If true, adds a learnable bias to the output. Default: true</param>
             /// <param name="device">The desired device of the parameters and buffers in this module</param>
             /// <param name="dtype">The desired floating point or complex dtype of the parameters and buffers in this module</param>
-            public static Conv3d Conv3d(long inputChannel, long outputChannel, long kernelSize, Padding padding, long stride = 1, long dilation = 1, PaddingModes paddingMode = PaddingModes.Zeros, long groups = 1, bool bias = true, Device? device = null, ScalarType? dtype = null)
+            public static Conv3d Conv3d(long in_channels, long out_channels, long kernelSize, Padding padding, long stride = 1, long dilation = 1, PaddingModes padding_mode = PaddingModes.Zeros, long groups = 1, bool bias = true, Device? device = null, ScalarType? dtype = null)
             {
-                var res = THSNN_Conv3d_ctor(inputChannel, outputChannel, kernelSize, stride, padding == Padding.Valid ? 0 : -1, dilation, (long)paddingMode, groups, bias, out var boxedHandle);
+                var res = THSNN_Conv3d_ctor(in_channels, out_channels, kernelSize, stride, padding == Padding.Valid ? 0 : -1, dilation, (long)padding_mode, groups, bias, out var boxedHandle);
                 if (res == IntPtr.Zero) { torch.CheckForErrors(); }
-                return new Conv3d(res, boxedHandle, inputChannel).MoveModule<Conv3d>(device, dtype);
+                return new Conv3d(res, boxedHandle, in_channels).MoveModule<Conv3d>(device, dtype);
             }
 
             /// <summary>
             /// Applies a 3D convolution over an input signal composed of several input planes
             /// </summary>
-            /// <param name="inputChannel">Number of channels in the input image</param>
-            /// <param name="outputChannel">Number of channels produced by the convolution</param>
+            /// <param name="in_channels">Number of channels in the input image</param>
+            /// <param name="out_channels">Number of channels produced by the convolution</param>
             /// <param name="kernelSize">Size of the convolving kernel</param>
             /// <param name="stride">Stride of the convolution. Default: (1,1,1)</param>
             /// <param name="padding">Zero-padding added to both sides of the input. padding=Valid is the same as no padding. padding=Same pads the input so the output has the shape as the input. </param>
             /// <param name="dilation">Spacing between kernel elements. Default: (1,1,1)</param>
-            /// <param name="paddingMode">'zeros', 'reflect', 'replicate' or 'circular'. Default: 'zeros'</param>
+            /// <param name="padding_mode">'zeros', 'reflect', 'replicate' or 'circular'. Default: 'zeros'</param>
             /// <param name="groups">Number of blocked connections from input channels to output channels. Default: 1</param>
             /// <param name="bias">If true, adds a learnable bias to the output. Default: true</param>
             /// <param name="device">The desired device of the parameters and buffers in this module</param>
             /// <param name="dtype">The desired floating point or complex dtype of the parameters and buffers in this module</param>
-            public static Conv3d Conv3d(long inputChannel, long outputChannel, (long, long, long) kernelSize, Padding padding, (long, long, long)? stride = null, (long, long, long)? dilation = null, PaddingModes paddingMode = PaddingModes.Zeros, long groups = 1, bool bias = true, Device? device = null, ScalarType? dtype = null)
+            public static Conv3d Conv3d(long in_channels, long out_channels, (long, long, long) kernelSize, Padding padding, (long, long, long)? stride = null, (long, long, long)? dilation = null, PaddingModes padding_mode = PaddingModes.Zeros, long groups = 1, bool bias = true, Device? device = null, ScalarType? dtype = null)
             {
                 if (stride == null) stride = (1, 1, 1);
                 if (dilation == null) dilation = (1, 1, 1);
 
-                var res = THSNN_Conv3d_ctor_1(inputChannel, outputChannel, kernelSize.Item1, kernelSize.Item2, kernelSize.Item3, stride.Value.Item1, stride.Value.Item2, stride.Value.Item3, padding == Padding.Valid ? 0 : -1, 0, 0, dilation.Value.Item1, dilation.Value.Item2, dilation.Value.Item3, (long)paddingMode, groups, bias, out var boxedHandle);
+                var res = THSNN_Conv3d_ctor_1(in_channels, out_channels, kernelSize.Item1, kernelSize.Item2, kernelSize.Item3, stride.Value.Item1, stride.Value.Item2, stride.Value.Item3, padding == Padding.Valid ? 0 : -1, 0, 0, dilation.Value.Item1, dilation.Value.Item2, dilation.Value.Item3, (long)padding_mode, groups, bias, out var boxedHandle);
                 if (res == IntPtr.Zero) { torch.CheckForErrors(); }
-                return new Conv3d(res, boxedHandle, inputChannel).MoveModule<Conv3d>(device, dtype);
+                return new Conv3d(res, boxedHandle, in_channels).MoveModule<Conv3d>(device, dtype);
             }
 
             public static partial class functional