+++ src/TorchSharp/NN/Convolution/ConvTranspose1D.cs
@@ -61,24 +61,24 @@ namespace TorchSharp
             /// <summary>
             /// Applies a 1D convolution over an input signal composed of several input planes.
             /// </summary>
-            /// <param name="inputChannel">Number of channels in the input image</param>
-            /// <param name="outputChannel">Number of channels produced by the convolution</param>
+            /// <param name="in_channels">Number of channels in the input image</param>
+            /// <param name="out_channels">Number of channels produced by the convolution</param>
             /// <param name="kernelSize">Size of the convolving kernel</param>
             /// <param name="stride">Stride of the convolution. Default: 1</param>
             /// <param name="padding">Zero-padding added to both sides of the input. Default: 0</param>
-            /// <param name="outputPadding">Additional size added to one side of the output shape. Default: 0</param>
+            /// <param name="output_padding">Additional size added to one side of the output shape. Default: 0</param>
             /// <param name="dilation">Spacing between kernel elements. Default: 1</param>
-            /// <param name="paddingMode">'zeros', 'reflect', 'replicate' or 'circular'. Default: 'zeros'</param>
+            /// <param name="padding_mode">'zeros', 'reflect', 'replicate' or 'circular'. Default: 'zeros'</param>
             /// <param name="groups">Number of blocked connections from input channels to output channels. Default: 1</param>
             /// <param name="bias">If true, adds a learnable bias to the output. Default: true</param>
             /// <param name="device">The desired device of the parameters and buffers in this module</param>
             /// <param name="dtype">The desired floating point or complex dtype of the parameters and buffers in this module</param>
             /// <returns>Tensor of shape (N,C_out,L_out)</returns>
-            public static ConvTranspose1d ConvTranspose1d(long inputChannel, long outputChannel, long kernelSize, long stride = 1, long padding = 0, long outputPadding = 0, long dilation = 1, PaddingModes paddingMode = PaddingModes.Zeros, long groups = 1, bool bias = true, Device? device = null, ScalarType? dtype = null)
+            public static ConvTranspose1d ConvTranspose1d(long in_channels, long out_channels, long kernelSize, long stride = 1, long padding = 0, long output_padding = 0, long dilation = 1, PaddingModes padding_mode = PaddingModes.Zeros, long groups = 1, bool bias = true, Device? device = null, ScalarType? dtype = null)
             {
-                var res = THSNN_ConvTranspose1d_ctor(inputChannel, outputChannel, kernelSize, stride, padding, outputPadding, dilation, (long)paddingMode, groups, bias, out var boxedHandle);
+                var res = THSNN_ConvTranspose1d_ctor(in_channels, out_channels, kernelSize, stride, padding, output_padding, dilation, (long)padding_mode, groups, bias, out var boxedHandle);
                 if (res == IntPtr.Zero) { torch.CheckForErrors(); }
-                return new ConvTranspose1d(res, boxedHandle, inputChannel).MoveModule<ConvTranspose1d>(device, dtype);
+                return new ConvTranspose1d(res, boxedHandle, in_channels).MoveModule<ConvTranspose1d>(device, dtype);
             }
 
             public static partial class functional
@@ -91,20 +91,20 @@ namespace TorchSharp
                 /// <param name="bias"></param>
                 /// <param name="stride"></param>
                 /// <param name="padding"></param>
-                /// <param name="outputPadding"></param>
+                /// <param name="output_padding"></param>
                 /// <param name="dilation"></param>
                 /// <param name="groups"></param>
                 /// <returns></returns>
                 public static Tensor conv_transpose1d(Tensor input, Tensor weight, Tensor? bias = null,
                     long? stride = null,
                     long? padding = null,
-                    long? outputPadding = null,
+                    long? output_padding = null,
                     long? dilation = null,
                     long groups = 1)
                 {
                     var strides = new long[] { stride ?? 1 };
                     var paddings = new long[] { padding ?? 0 };
-                    var outputPaddings = new long[] { outputPadding ?? 0 };
+                    var outputPaddings = new long[] { output_padding ?? 0 };
                     var dilations = new long[] { dilation ?? 1 };
                     var biasHandle = (bias is null ? IntPtr.Zero : bias.Handle);
                     unsafe {